Complete re-write of the GUI, moving some soundsystem init stuff to GUI script.
